requirements.txt
Django>=4.2.4,<4.3
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.26.4,<0.27
djangorestframework-simplejwt>=5.2.2,<5.3
requests>=2.31.0,<2.32
graphene-django>=3.1.5<3.2
python-dotenv>=1.0.0,<1.1